Refactor xml into root and disperse const values to relevant modules
