+++ epp-client/README.md
@@ -56,7 +56,7 @@ use std::collections::HashMap;
 use epp_client::config::{EppClientConfig, EppClientConnection};
 use epp_client::EppClient;
 use epp_client::epp::{EppDomainCheck, EppDomainCheckResponse};
-use epp_client::epp::generate_client_tr_id;
+use epp_client::generate_client_tr_id;
 
 #[tokio::main]
 async fn main() {
