+++ epp-client/src/contact.rs
@@ -3,3 +3,5 @@ pub mod create;
 pub mod delete;
 pub mod info;
 pub mod update;
+
+pub const EPP_CONTACT_XMLNS: &str = "urn:ietf:params:xml:ns:contact-1.0";
