+++ epp-client/src/domain.rs
@@ -6,3 +6,5 @@ pub mod renew;
 pub mod rgp;
 pub mod transfer;
 pub mod update;
+
+pub const EPP_DOMAIN_XMLNS: &str = "urn:ietf:params:xml:ns:domain-1.0";
