+++ epp-client/src/domain/check.rs
@@ -2,8 +2,8 @@
 
 use epp_client_macros::*;
 
+use super::EPP_DOMAIN_XMLNS;
 use crate::common::{ElementName, EppObject, StringValue};
-use crate::epp::xml::EPP_DOMAIN_XMLNS;
 use crate::request::Command;
 use crate::response::CommandResponse;
 use serde::{Deserialize, Serialize};
@@ -18,7 +18,7 @@ use serde::{Deserialize, Serialize};
 /// use epp_client::config::{EppClientConfig, EppClientConnection};
 /// use epp_client::EppClient;
 /// use epp_client::domain::check::{EppDomainCheck, EppDomainCheckResponse};
-/// use epp_client::epp::generate_client_tr_id;
+/// use epp_client::generate_client_tr_id;
 ///
 /// #[tokio::main]
 /// async fn main() {
