+++ epp-client/src/domain/create.rs
@@ -2,11 +2,11 @@
 
 use epp_client_macros::*;
 
+use super::EPP_DOMAIN_XMLNS;
 use crate::common::{
     DomainAuthInfo, DomainContact, ElementName, EppObject, HostAttr, HostAttrList, HostList,
     HostObjList, Period, StringValue,
 };
-use crate::epp::xml::EPP_DOMAIN_XMLNS;
 use crate::request::Command;
 use crate::response::CommandResponse;
 use serde::{Deserialize, Serialize};
@@ -23,7 +23,7 @@ use serde::{Deserialize, Serialize};
 /// use epp_client::EppClient;
 /// use epp_client::common::DomainContact;
 /// use epp_client::domain::create::{EppDomainCreate, EppDomainCreateResponse};
-/// use epp_client::epp::generate_client_tr_id;
+/// use epp_client::generate_client_tr_id;
 ///
 /// #[tokio::main]
 /// async fn main() {
