+++ epp-client/src/domain/rgp.rs
@@ -2,3 +2,5 @@
 
 pub mod report;
 pub mod request;
+
+pub const EPP_DOMAIN_RGP_EXT_XMLNS: &str = "urn:ietf:params:xml:ns:rgp-1.0";
