+++ epp-client/src/domain/rgp/request.rs
@@ -3,8 +3,9 @@
 use epp_client_macros::*;
 
 use crate::common::{ElementName, EmptyTag, EppObject, Extension};
+use crate::domain::rgp::EPP_DOMAIN_RGP_EXT_XMLNS;
 use crate::domain::update::{DomainChangeInfo, DomainUpdateRequest, DomainUpdateRequestData};
-use crate::epp::xml::{EPP_DOMAIN_RGP_EXT_XMLNS, EPP_DOMAIN_XMLNS};
+use crate::domain::EPP_DOMAIN_XMLNS;
 use crate::request::CommandWithExtension;
 use crate::response::CommandResponseWithExtension;
 use serde::{Deserialize, Serialize};
@@ -19,7 +20,7 @@ use serde::{Deserialize, Serialize};
 /// use epp_client::config::{EppClientConfig, EppClientConnection};
 /// use epp_client::EppClient;
 /// use epp_client::domain::rgp::request::{EppDomainRgpRestoreRequest, EppDomainRgpRestoreRequestResponse};
-/// use epp_client::epp::generate_client_tr_id;
+/// use epp_client::generate_client_tr_id;
 ///
 /// #[tokio::main]
 /// async fn main() {
