+++ epp-client/src/domain/update.rs
@@ -6,7 +6,7 @@ use crate::common::{
     DomainAuthInfo, DomainContact, DomainStatus, ElementName, EppObject, HostList, StringValue,
 };
 
-use crate::epp::xml::EPP_DOMAIN_XMLNS;
+use super::EPP_DOMAIN_XMLNS;
 use crate::request::Command;
 use crate::response::EppCommandResponse;
 use serde::{Deserialize, Serialize};
@@ -23,7 +23,7 @@ use serde::{Deserialize, Serialize};
 /// use epp_client::EppClient;
 /// use epp_client::common::{DomainStatus, DomainContact};
 /// use epp_client::domain::update::{EppDomainUpdate, EppDomainUpdateResponse, DomainAddRemove};
-/// use epp_client::epp::generate_client_tr_id;
+/// use epp_client::generate_client_tr_id;
 ///
 /// #[tokio::main]
 /// async fn main() {
