--- epp-client/src/epp.rs
@@ -1,5 +1,0 @@
-//! Types for EPP requests and responses
-
-pub mod xml;
-
-pub use crate::connection::client::default_client_tr_id_fn as generate_client_tr_id;
