+++ epp-client/src/host.rs
@@ -3,3 +3,5 @@ pub mod create;
 pub mod delete;
 pub mod info;
 pub mod update;
+
+pub const EPP_HOST_XMLNS: &str = "urn:ietf:params:xml:ns:host-1.0";
