+++ epp-client/src/host/update.rs
@@ -3,7 +3,7 @@
 use epp_client_macros::*;
 
 use crate::common::{ElementName, EppObject, HostAddr, HostStatus, StringValue};
-use crate::epp::xml::EPP_HOST_XMLNS;
+use crate::host::EPP_HOST_XMLNS;
 use crate::request::Command;
 use crate::response::EppCommandResponse;
 use serde::{Deserialize, Serialize};
@@ -19,7 +19,7 @@ use serde::{Deserialize, Serialize};
 /// use epp_client::EppClient;
 /// use epp_client::common::{HostAddr, HostStatus};
 /// use epp_client::host::update::{EppHostUpdate, EppHostUpdateResponse, HostAddRemove, HostChangeInfo};
-/// use epp_client::epp::generate_client_tr_id;
+/// use epp_client::generate_client_tr_id;
 ///
 /// #[tokio::main]
 /// async fn main() {
